rename current login to form and use modal for transition, update references
