+++ web/src/components/LoginForm/LoginForm.stories.tsx
@@ -1,0 +1,12 @@
+import type { ComponentMeta } from '@storybook/react'
+
+import LoginForm from './LoginForm'
+
+export const generated = () => {
+  return <LoginForm />
+}
+
+export default {
+  title: 'Pages/LoginForm',
+  component: LoginForm,
+} as ComponentMeta<typeof LoginForm>
