+++ web/src/components/LoginModal/LoginModal.stories.tsx
@@ -1,12 +1,25 @@
+// When you've added props to your component,
+// pass Storybook's `args` through this story to control it from the addons panel:
+//
+// ```tsx
+// import type { ComponentStory } from '@storybook/react'
+//
+// export const generated: ComponentStory<typeof LoginModal> = (args) => {
+//   return <LoginModal {...args} />
+// }
+// ```
+//
+// See https://storybook.js.org/docs/react/writing-stories/args.
+
 import type { ComponentMeta } from '@storybook/react'
 
-import LoginPage from './LoginModal'
+import LoginModal from './LoginModal'
 
 export const generated = () => {
-  return <LoginPage />
+  return <LoginModal />
 }
 
 export default {
-  title: 'Pages/LoginPage',
-  component: LoginPage,
-} as ComponentMeta<typeof LoginPage>
+  title: 'Components/LoginModal',
+  component: LoginModal,
+} as ComponentMeta<typeof LoginModal>
